LayoutTests/media/track/track-default-attribute-expected.txt

Tests that a track with the 'default' attribute loads automatically.

EVENT(load)
EXPECTED (event.target.readyState == '2') OK
EXPECTED (event.target.id == 'default') OK

END OF TEST
